Refactor training setup
